// src/views/detail/djradio-detail/store/type.ts
export type IradioDetailType = {
  id: number
  name: string
  dj: {
    defaultAvatar: boolean
    province: number
    authStatus: number
    followed: boolean
    avatarUrl: string
    accountStatus: number
    gender: number
    city: number
    birthday: number
    userId: number
    userType: number
    nickname: string
    signature: string
    description: string
    detailDescription: string
    avatarImgId: number
    backgroundImgId: number
    backgroundUrl: string
    authority: number
    mutual: boolean
    expertTags: any
    experts: any
    djStatus: number
    vipType: number
    remarkName: any
    authenticationTypes: number
    avatarDetail: any
    avatarImgIdStr: string
    backgroundImgIdStr: string
    anchor: boolean
    rewardCount: number
    avatarImgId_str: string
    canReward: boolean
  }
  picId: number
  picUrl: string
  desc: string
  subCount: number
  shareCount: number
  likedCount: number
  programCount: number
  commentCount: number
  createTime: number
  categoryId: number
  category: string
  secondCategoryId: number
  secondCategory: string
  radioFeeType: number
  feeScope: number
  lastProgramCreateTime: number
  lastProgramId: number
  rcmdText: string
  subed: boolean
  commentDatas: Array<any>
  feeInfo: any
  unlockInfo: any
  original: boolean
  playCount: number
  privacy: boolean
  disableShare: boolean
  icon: any
  activityInfo: any
  toplistInfo: any
  dynamic: boolean
  labelDto: any
  labels: any
  detailRcmdTabOrpheus: string
  toast: any
  limitFreeToast: any
}
// ...
export type IradioDetailState = {
  radioDetail: IradioDetailType
  radioProgram: radioProgramType
}